+++ app/src/main/java/com/example/ttapp/viewmodel/SurveyViewModel.java
@@ -1,0 +1,5 @@
+package com.example.ttapp.viewmodel;
+
+public class SurveyViewModel {
+
+}
